Implement the framework for adding in a recursive search disassembler.
This commit introduces a new enumeration for selecting the algorithm,
the skeleton class for the new disassembler and two new command line
options to select between the different algorithms.

The two new command line options are '--linear' and '--recursive' to
select between the linear sweep and recursive search algorithms. By
default, the linear sweep algorithm is activated.